Simplifed permission validation.
