+++ backend/src/middlewares/validatePermission.js
@@ -1,57 +1,34 @@
 const HttpStatus = require('http-status-codes');
 const log = require('../components/logger');
 const { getRoles } = require('../components/lookup');
-const { PERMISSIONS } = require('../util/constants');
 
-module.exports = {
-  /**
-   * Validates that the user has the specified permission.
-   * @param {*} requiredPermissions
-   * @returns
-   */
-  validateUserHasPermissions(...requiredPermissions) {
-    return async function (req, res, next) {
-      log.verbose(`validating permissions ${requiredPermissions}`);
+/**
+ * Validates that the user has the specified permission.
+ * @param {*} requiredPermissions
+ * @returns
+ */
+module.exports = function (...requiredPermissions) {
+  return async function (req, res, next) {
+    log.verbose(`validating permissions ${requiredPermissions}`);
 
-      // Reject deactivated users
-      if (req.session?.passport?.user?.statecode === 1) {
-        log.info('User is deactivated, responding with 401');
-        return res.status(HttpStatus.UNAUTHORIZED).json();
-      }
+    // Reject deactivated users
+    if (req.session?.passport?.user?.statecode === 1) {
+      log.info('User is deactivated, responding with 401');
+      return res.status(HttpStatus.UNAUTHORIZED).json();
+    }
 
-      const userRole = req.session?.passport?.user?.role;
+    const userRole = req.session?.passport?.user?.role;
 
-      if (!userRole) {
-        return res.sendStatus(HttpStatus.FORBIDDEN);
-      }
-
-      const roles = await getRoles();
-      const matchingRole = roles.find((role) => role.data.roleNumber === userRole.roleNumber);
-
-      const permissions = matchingRole ? matchingRole.data.permissions?.map((p) => p.permissionNumber) : [];
-
-      const valid = requiredPermissions?.some((p) => permissions.includes(p));
-
-      valid ? next() : res.sendStatus(HttpStatus.FORBIDDEN);
-    };
-  },
-  async validateUserCanEditOther(req, res, next) {
-    const currentUser = req.session?.passport?.user;
-    const targetContactId = req.params.contactId;
-    log.verbose('Checking if user can edit self or others');
+    if (!userRole) {
+      return res.sendStatus(HttpStatus.FORBIDDEN);
+    }
 
     const roles = await getRoles();
-    const matchingRole = roles.find((role) => role.data.roleNumber === currentUser.role.roleNumber);
-    const permissions = matchingRole?.data?.permissions?.map((p) => p.permissionNumber) || [];
+    const matchingRole = roles.find((role) => role.data.roleNumber === userRole.roleNumber);
+    const permissions = matchingRole ? matchingRole.data.permissions?.map((p) => p.permissionNumber) : [];
 
-    const canEditUsers = permissions.includes(PERMISSIONS.EDIT_USERS);
-    const canUpdateSelf = permissions.includes(PERMISSIONS.UPDATE_SELF);
+    const valid = requiredPermissions?.some((p) => permissions.includes(p));
 
-    if (canEditUsers) return next();
-
-    if (canUpdateSelf && currentUser.contactId === targetContactId) {
-      return next();
-    }
-    return res.sendStatus(HttpStatus.FORBIDDEN);
-  },
+    valid ? next() : res.sendStatus(HttpStatus.FORBIDDEN);
+  };
 };
