+++ backend/src/routes/organization.js
@@ -2,7 +2,7 @@ const express = require('express');
 const passport = require('passport');
 const router = express.Router();
 const auth = require('../components/auth');
-const { validateUserHasPermissions } = require('../middlewares/validatePermission');
+const validatePermission = require('../middlewares/validatePermission');
 const isValidBackendToken = auth.isValidBackendToken();
 const { createOrganization, getOrganization, getOrganizationFacilities, updateOrganization } = require('../components/organization');
 const { PERMISSIONS } = require('../util/constants');
@@ -79,7 +79,7 @@ router.put(
   '/:organizationId',
   passport.authenticate('jwt', { session: false }),
   isValidBackendToken,
-  validateUserHasPermissions(PERMISSIONS.CHANGE_ORG_INFORMATION),
+  validatePermission(PERMISSIONS.CHANGE_ORG_INFORMATION),
   [(param('organizationId', 'URL param: [organizationId] is required').notEmpty().isUUID(), checkSchema(organizationSchema))],
   (req, res) => {
     validationResult(req).throw();
